# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farmácia Pai e Filha API",
        "description": "API para gerenciamento da Farmácia Pai e Filha",
        "version": "1.0.0"
    },
    "paths": {
...
        "/product/alerts/stock": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retorna JSON de produtos com estoque baixo ou zerado",
                "operationId": "626a4a51bf8ce31f028b260d5e539f7c",
                "responses": {
                    "200": {
                        "description": "Lista de alertas em JSON"
                    },
                    "403": {
                        "description": "Não autorizado"
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Exibe os detalhes de um produto específico",
                "operationId": "f2f7fd834f846a9529e8a7b87d474cb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a view de detalhes do produto"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        },
        "/product/all-products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Lista todos os produtos com paginação e ordenação",
                "operationId": "0b560a752c43a4ca0ba6c998088c0e03",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de itens (10, 15, 20, 25, 30)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna a view do estoque de produtos"
                    }
                }
            }
        },
...
        "/create-product": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cria um produto e opcionalmente fornecedor/lote",
                "description": "Registra o produto e, se solicitado, cria o fornecedor com endereço e o primeiro lote.",
                "operationId": "450a0d405226a5c2a1305695ed1bb3b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idCategory",
                                    "name",
                                    "price",
                                    "barcode"
                                ],
                                "properties": {
                                    "idCategory": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Dipirona 500mg"
                                    },
                                    "barcode": {
                                        "type": "string",
                                        "example": "7891234567890"
                                    },
                                    "price": {
                                        "type": "string",
                                        "example": "15,50"
                                    },
                                    "min_stock_alert": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "novoFornecedor": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "12345678000199"
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "example": "41000000"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Avenida Sete de Setembro"
                                    },
                                    "number": {
                                        "type": "string",
                                        "example": "100"
                                    },
                                    "neighborhood": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Salvador"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "BA"
                                    },
                                    "batch_code": {
                                        "type": "string",
                                        "example": "LOT-2026-X"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "expiration_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2028-12-31"
                                    },
                                    "cost_price": {
                                        "type": "string",
                                        "example": "5,50"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    },
                    "500": {
                        "description": "Erro Interno"
                    }
                }
            }
        },
        "/product/update/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualiza produto e lote simultaneamente",
                "operationId": "9b072fa387b78c4661dbe7a637bdbac4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "idCategory": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "string"
                                    },
                                    "image_product": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "quantity_now": {
                                        "type": "integer"
                                    },
                                    "idBatch": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        },
        "/product/deactivate/{id}": {
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Inativa ou reativa um produto (Soft Delete)",
                "operationId": "aedfd8ea60fa0c4118ac7be1faa89ce6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status alterado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        },
...
    },
...
}